*Really* support models that are nested inside modules.
The previous version only supported "leaf" models that
are nested inside modules. With this commit it is now
possible to use models that are nested inside modules
all the way down the nesting strategy. So, on a
"/articles/1/comments/1/ratings/1" route, you will be
able to use classes Article, Community::Comment and
Community::Rating (previously only Community::Rating
was supported).